+++ CHANGELOG.md
@@ -6,10 +6,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2019-09-6
+### Added
+- callback to each dnd event
+
+### Updated
+- Example #4 page
+
+### Removed
+- auto invoke init function (now the user needs to call the plugin init function)
+
 ## [1.1.0] - 2019-08-20
 ### Added
 - new method for making droppable or draggable any element not just the default ones
 - some code corrections 
 
-[Unreleased]: https://github.com/2rhop/css-dnd/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/2rhop/css-dnd/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/2rhop/css-dnd/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/2rhop/css-dnd/compare/v0.1.0...v1.1.0